+++ src/components/app/Overview/Overview.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useMemo, useState } from 'react'
 import moment from 'moment'
-import { Link, useParams } from 'react-router-dom'
+import { Link, useHistory, useLocation, useParams } from 'react-router-dom'
 import { toast } from 'react-toastify'
-import { ModuleNameMap, Moment12HourFormat, OVERVIEW_TABS, URLS } from '../../../config'
+import { ModuleNameMap, Moment12HourFormat, URLS } from '../../../config'
 import { getJobCIPipeline, getTeamList } from '../../../services/service'
 import {
     showError,
@@ -33,12 +33,19 @@ import { getAppConfig, getGitProviderIcon } from './utils'
 import { EnvironmentList } from './EnvironmentList'
 import { MAX_LENGTH_350 } from '../../../config/constantMessaging'
 import { getModuleInfo } from '../../v2/devtronStackManager/DevtronStackManager.service'
+import { OVERVIEW_TABS, TAB_SEARCH_KEY } from './constants'
 const MandatoryTagWarning = importComponentFromFELibrary('MandatoryTagWarning')
 const Catalog = importComponentFromFELibrary('Catalog')
 const DependencyList = importComponentFromFELibrary('DependencyList')
 
+type AvailableTabs = typeof OVERVIEW_TABS[keyof typeof OVERVIEW_TABS]
+
 export default function AppOverview({ appMetaInfo, getAppMetaInfoRes, filteredEnvIds, appType }: AppOverviewProps) {
     const { appId: appIdFromParams } = useParams<{ appId: string }>()
+    const location = useLocation();
+    const history = useHistory()
+    const searchParams = new URLSearchParams(location.search)
+    const activeTab = searchParams.get(TAB_SEARCH_KEY) as AvailableTabs
     const config = getAppConfig(appType)
     const isJobOverview = appType === 'job'
     const isHelmChart = appType === 'helm-chart'
@@ -55,7 +62,6 @@ export default function AppOverview({ appMetaInfo, getAppMetaInfoRes, filteredEn
     const [newUpdatedBy, setNewUpdatedBy] = useState<string>()
     const [jobPipelines, setJobPipelines] = useState<JobPipeline[]>([])
     const [reloadMandatoryProjects, setReloadMandatoryProjects] = useState<boolean>(true)
-    const [activeTab, setActiveTab] = useState<typeof OVERVIEW_TABS[keyof typeof OVERVIEW_TABS]>(OVERVIEW_TABS.ABOUT)
     const [, isArgoInstalled] = useAsync(
         () => getModuleInfo(ModuleNameMap.ARGO_CD),
         [],
@@ -64,6 +70,19 @@ export default function AppOverview({ appMetaInfo, getAppMetaInfoRes, filteredEn
 
     let _moment: moment.Moment
     let _date: string
+
+    const setActiveTab = (selectedTab: AvailableTabs) => {
+        searchParams.set(TAB_SEARCH_KEY, selectedTab)
+        history.replace({ search: searchParams.toString() })
+    }
+
+    useEffect(() => {
+        // add a default tab if not set
+        if (!activeTab || !Object.values(OVERVIEW_TABS).includes(activeTab)) {
+            searchParams.set(TAB_SEARCH_KEY, OVERVIEW_TABS.ABOUT)
+            history.replace({ search: searchParams.toString() })
+        }
+    }, [searchParams])
 
     useEffect(() => {
         if (appMetaInfo?.appName) {
@@ -455,7 +474,7 @@ export default function AppOverview({ appMetaInfo, getAppMetaInfoRes, filteredEn
                     <RadioGroup
                         className="gui-yaml-switch gui-yaml-switch--lg gui-yaml-switch-window-bg flex-justify-start dc__no-background-imp"
                         name="overview-tabs"
-                        initialTab={OVERVIEW_TABS.ABOUT}
+                        initialTab={activeTab}
                         disabled={false}
                         onChange={(e) => {
                             setActiveTab(e.target.value)
