+++ src/components/app/Overview/constants.ts
@@ -10,3 +10,15 @@ export const loadingEnvironmentList: AppEnvironment[] = Array.from(Array(3).keys
     infraMetrics: false,
     prod: false,
 }))
+
+/**
+ * Tabs for the overview of the app(s)
+ */
+export const OVERVIEW_TABS = {
+    ABOUT: 'about',
+    ENVIRONMENTS: 'environments',
+    JOB_PIPELINES: 'job-pipelines',
+    DEPENDENCIES: 'dependencies',
+} as const
+
+export const TAB_SEARCH_KEY = 'tab'
