+++ src/config/constants.ts
@@ -790,13 +790,3 @@ export const SERVER_ERROR_CODES = {
 }
 
 export const ENV_ALREADY_EXIST_ERROR = 'Deployment pipeline already exists for this environment'
-
-/**
- * Tabs for the overview of the app(s)
- */
-export const OVERVIEW_TABS = {
-    ABOUT: 'about',
-    ENVIRONMENTS: 'environments',
-    JOB_PIPELINES: 'job-pipelines',
-    DEPENDENCIES: 'dependencies',
-} as const
